requirements.txt
bcrypt==4.0.1
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
DateTime==5.1
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.29.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.3
WTForms==3.0.1
zope.interface==6.0
